virts-509-update agent executors on heartbeat (#593)
* virts-509-update agent executors on heartbeat
If agent heartbeat now has changed executors and/or priorities for
executors, handle_heartbeat function now updates the executors in the
database.

* virts-509-update agent executors on heartbeat
    If agent heartbeat now has changed executors and/or preference for
    executors, handle_heartbeat function now updates the executors in the
    database utilizing update_agent function defined in data_svc.py.

* virts-509-update agent executors on heartbeat
reworked data_svc function, now only removes executor from database if
it is not in the new and incoming executors from heartbeat. If new
executor is not in database, it is also added.

* virts-509-update agent executors on heartbeat
Instead of changing the incoming executors list into dictionaries with
preferred keys/values, list of dictionaries is now converted into an
ordered list where first item is given preference. This allows for set
operations.

* virts-509-update agent executors on heartbeat
Utilizing the update function. Update function should be updated to be
able to more easily intake various keys for granular control of which
items need to be updated especially if they share agent_ids

* virts-509-update agent executors on heartbeat
Modified conf/core.sql to make agent_id,executor combinations
unique, now when we add an agent_id/executor combination to the table,
if it already exists it replaces it and thus updates the preferred
field.