Allow configuring which report fields to redact
This makes it easier to change what fields get redacted from reports.

Keys in the configuration are parsed to traverse the report to where the
fields to be redacted are located. '[]' indicates a list, '{}' indicates
an dictionary, and '.' indicates a property. For example,
'adversary.phases{}[]' means that in the report, under the key
'adversary', there is the 'phases' property which holds a dictionary,
the values of which are lists containing the dictionaries that have the
actual fields to be redacted.

Keys in the config must start with the name of one of the top-level
properties in the report. A side effect of this config handling is that
no keys in the report that need to be redacted can contain the '.'
character.