Make it easier to get redacted report
If `op` is an Operation, you can now access the redacted report with:
    op.get_report(True)
instead of
    Operation.report.fget(op, redacted=True)

This change is necessary because @properties aren't supposed to have
parameters. The alternative is to just use `get_report` and not have a
property at all, but I left it as a property for backwards
compatibility.