Allow adding adversary packs in GUI
Phases from the pack can be merged into existing phases or inserted as
new ones. If merging, it doesn't matter which phase's "add pack" button
you select. If not merging, the phases will be added after the one whose
"add pack" button was selected.

When the adversary is saved the abilities from the pack will be added.
The resulting adversary YAML will have the ids of all the abilities in
the pack, not just a single entry with the pack id.