Don't check cleanup links against requirements
We don't need to check requirements for cleanup links. Without this fix,
some cleanup links won't get run because the requirement on the ability
is no longer valid. For example, an ability with a requirement that a
file doesn't exist which creates that file. When the cleanup link was
checked against the requirement it would fail because the ability had
created it.