Fixed issue with command variable substition
If a command contained variables for a, a.b, and a.b.c,
when substituting 'a' with a value, it had been filling in
values for a, a.b, and a.b.c due to improper regex matching.

Also fixed an issue where characters in a variable/trait name
would be treated as regex special characters (e.g., dots).