Possible fix to Issue #2315 (affects `templates/abilities.html`)
The REST API expects an ability object in which the cleanup field is an array (list) of strings. Currently, the "Create Ability" UI saves the cleanup command as a string, causing an Unprocessable Entity response status code (422) when executing the PUT request in API v2.  This commit includes a slight change to the cleanup field in the Create Ability modal which converts a line-break separated list of cleanup commands to the correct type.  The net effect of the change is to enable users to create and edit abilities which include cleanup commands through the UI and API v2.