Change addPotentialLink to have ability: link in response.
Currently link is unpacked into the response, but in the operation_api_manager.py create_potential_link which handles this request it is expected that the data has an ability key. Since there is none, when it goes to build an ability it builds a default 'manual' one instead.

This can be tested by making an operation, and then adding a potential link.