Creates a new operation per schedule
When an object schedule is created, it embeds a complete operation
template in its "task" field.
On the first schedule time, it creates an operation from the template.
Until now, this operation was used for each subsequent schedule.
This commit prevent that and creates a new operation per schedule, with
the start date added in its name.