fix: docker-compose, volumes don't overwrite UI
currently, mounting all of the working directory to the app dir in
the container means that the UI built in the container at
`/usr/src/app/plugins/magma` is overwritten by what is on the host.
If the UI is not built on the host, the container and docker-compose
will not work as expected.

This change adds a volume to protect the contents of the built UI in
the container from being overwritten by the host's content in
`/usr/src/app/plugins/magma`